refactor(P3-3): WbsAnalyticsHandler を Application Service に改修(②案)
- QueryBus 未登録で「DIから直接取得して独自メソッドを呼ぶ」なんちゃって CQRS を解消
- IWbsAnalyticsApplicationService を定義し WbsAnalyticsApplicationService にリネーム
  （query/ から applications/wbs/ へ移動）。docs/08「迷ったら CQRS を使わない」に沿う
- SYMBOL.WbsAnalyticsHandler → SYMBOL.IWbsAnalyticsApplicationService、DI バインドと
  actions.ts の解決を IF 経由に更新。テストもリネームし new を差し替え

本番型チェックエラーなし、該当テスト 8 件 green、eslint クリーン。
